refactor(core): decouple FeedbackRepository from SQLite (Phase 2)
