+++ packages/memento-core/src/domains/memory/repositories/feedback-repository.interface.ts
@@ -1,0 +1,7 @@
+import type { CreateFeedbackEventInput, FeedbackNetScore } from '../../../shared/types/feedback.types.js';
+
+export interface IFeedbackRepository {
+  insertFeedback(input: CreateFeedbackEventInput): { id: number; created_at: string };
+  getNetScores(memoryIds: string[], windowDays?: number): Map<string, number>;
+  getNetScoreRows(memoryIds: string[], windowDays?: number): FeedbackNetScore[];
+}
